server/requirements.txt
Flask
pymongo
flask_cors  # need when gets cors erros
dnspython
pyjwt
flask-jwt-extended
opencv-contrib-python
mediapipe
Pillow